+++ src/ebfm/coupling/couplers/yacCoupler.py
@@ -13,7 +13,7 @@ from .base import Coupler, Grid, Dict, CouplingConfig, CouplerErrorCode
 from ebfm.coupling.fields import FieldSet
 from ebfm.coupling.fields import YACField as Field
 
-from typing import Union, Tuple, Optional
+from typing import Tuple, Optional
 
 # from ebfm.geometry import Grid  # TODO: consider introducing a new data structure native to EBFM?
 
@@ -42,7 +42,7 @@ class YACCoupler(Coupler):
         self.grid: yac.UnstructuredGrid = None
         self.corner_points: yac.Points = None
 
-    def setup(self, grid: Union[Dict, Grid], time: Dict[str, float]):
+    def setup(self, grid: Dict | Grid, time: Dict[str, float]):
         """
         Setup the coupling interface